+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Embedded/EmbeddedPaymentElement+Internal.swift
@@ -191,7 +191,8 @@ extension EmbeddedPaymentElement: EmbeddedPaymentMethodsViewDelegate {
                                                                                canUpdate: elementsSession.paymentMethodUpdateForPaymentSheet,
                                                                                isCBCEligible: paymentMethod.isCoBrandedCard && elementsSession.isCardBrandChoiceEligible,
                                                                                allowsSetAsDefaultPM: elementsSession.paymentMethodSetAsDefaultForPaymentSheet,
-                                                                               isDefault: paymentMethod == defaultPaymentMethod)
+                                                                               isDefault: paymentMethod == defaultPaymentMethod,
+                                                                               defaultBillingDetails: configuration.defaultBillingDetails)
             let removeSavedPaymentMethodMessage = UpdatePaymentMethodViewController.resolveRemoveMessage(
                 removeSavedPaymentMethodMessage: configuration.removeSavedPaymentMethodMessage,
                 paymentMethodRemoveIsPartial: elementsSession.paymentMethodRemoveIsPartialForPaymentSheet(),